Update script.js
Updated form creation to use a new form created by the user running the script to reduce complexity and permission issues. Updated prompts at end of form creation to account for opening to edit form's style or simply opening the shareable Form. Added a dropdown function to open the form.