Ignore completely empty responses on Weekly Sheets
Updated to include a "counta" function dropped in column B for the weekly sheet that checks to see if there are any imported picks for that member, otherwise it will leave it blank. Columns C and D rely on a value in B and therefore were modified to not calculate unless there is a value in column B of the same row.