Fix % Correct formula, data restore, minor changes
Issue found by @AustinOrphan with the formula that was only noticeable when only Thursday picks were imported. 

Also fixed an issue with the total correct formula.

Data restore function was failing to bring in the final data row due to missing a "<=" in a loop that terminated one row earlier with a "<". 

Removed unnecessary "Logger.log" testing lines.